backend/cachingsystem/requirements.txt
fastapi
uvicorn
pydantic
requests
python-jose